pyproject.toml
[project]
name = "exam-eye"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mediapipe==0.10.7",
    "numpy==1.24.3",
    "opencv-python==4.8.1.78",
    "python-dotenv==1.0.0",
    "ultralytics==8.3.0",
]
